+++ src/System.Buffers.Primitives/System/Internal/ManagedBufferPool.cs
@@ -3,7 +3,7 @@
 
 namespace System.Buffers.Internal
 {
-    internal sealed partial class ManagedBufferPool : BufferPool
+    internal sealed class ManagedBufferPool : BufferPool
     {
         readonly static ManagedBufferPool s_shared = new ManagedBufferPool();
 
@@ -12,6 +12,29 @@ namespace System.Buffers.Internal
             get
             {
                 return s_shared;
+            }
+        }
+
+        public override OwnedBuffer<byte> Rent(int minimumBufferSize)
+        {
+            return new ArrayPoolMemory(minimumBufferSize);
+        }
+
+        protected override void Dispose(bool disposing)
+        {
+        }
+
+        private class ArrayPoolMemory : OwnedBuffer<byte>
+        {
+            public ArrayPoolMemory(int size) : base(ArrayPool<byte>.Shared.Rent(size))
+            {
+            }
+
+            protected override void Dispose(bool disposing)
+            {
+                ArrayPool<byte>.Shared.Return(Array);
+
+                base.Dispose(disposing);
             }
         }
     }
